python/requirements.txt
scikit-learn>=1.0
joblib>=1.0
mysql-connector-python>=8.0
numpy>=1.21
pandas>=1.3
